requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
Jinja2
fpdf
